initial cleanup of file structure
